requirements.txt
pandas
streamlit
plotly
python-dotenv
requests
